+++ packages/common/test/serialization.test.ts
@@ -171,5 +171,7 @@ test('serialize type value and serializeUpdateContractParameters give same resul
 });
 
 test('serializeTypeValue throws an error if unable to serialize', () => {
-    expect(() => serializeTypeValue('test', Buffer.alloc(0))).toThrow();
+    expect(() => serializeTypeValue('test', Buffer.alloc(0))).toThrowError(
+        Error
+    );
 });